feat(exerciseAPI) : add API호출, DB insert code for Execise

+++ backend/src/main/java/org/synergym/backend/service/PostService.java
@@ -29,7 +29,7 @@ public interface PostService {
                 .title(post.getTitle())
                 .content(post.getContent())
                 .viewCount(post.getViewCount())
-                .userId(post.getWriter().getId().longValue())
+                .userId(post.getWriter().getId())
                 .build();
     }
